+++ CHANGELOG.md
@@ -9,7 +9,9 @@ Master
 * BC BREAK: Refactored ExclusionStrategy interface
 * FIXED: Infinite loop on circular references when excluding property
 * ADDED: Type Adapter caching
+* ADDED: Support for subclasses of DateTime
 * REMOVED: ArrayList/HashMap type adapters
+* REMOVED: Support for 'JsonElement' type, must use full class name
 
 v0.1.0
 ------